+++ projects/components/src/lib/button/button.directive.ts
@@ -5,11 +5,29 @@ import { Directive, ElementRef, Input, Renderer2, OnChanges, SimpleChanges, Host
   selector: '[MasButton]', 
 })
 export class MasButtonDirective implements OnChanges{
+  /**
+   * The variant of the button including 'primary' | 'secondary' | 'tertiary' | 'accent'.
+   */
   @Input() variant: 'primary' | 'secondary' | 'tertiary' | 'accent' = 'primary';
+  /**
+   * The size of the button. Available options: 'small', 'medium'. Default: 'medium'.
+   */
   @Input() size: 'small' | 'medium' = 'medium';
+  /**
+   * Whether the button should be disabled. Default: false.
+   */
   @Input() disabled: boolean = false;
+  /**
+   * The name of the icon to display on the left side of the button. 
+   */
   @Input() iconLeft: string | null = null;
+  /**
+   * The name of the icon to display on the right side of the button.
+   */
   @Input() iconRight: string | null = null;
+  /**
+   * The name of the icon to display as the sole content of the button.
+   */
   @Input() iconAlone: string | null = null;
 
   @HostBinding('class.mas-btn') vtmnBtnClass = true;